Move dedicated tooltip styling
