+++ packages/ui/src/common/components/icons/index.ts
@@ -2,6 +2,7 @@ export * from './ArrowDownExpandedIcon'
 export * from './ArrowIcon'
 export * from './ArrowInsideIcon'
 export * from './ArrowOutsideIcon'
+export * from './ArrowRightIcon'
 export * from './BlockIcon'
 export * from './CheckboxIcon'
 export * from './CopyIcon'
@@ -9,6 +10,8 @@ export * from './CrossIcon'
 export * from './FailureIcon'
 export * from './FounderMemberIcon'
 export * from './Icon'
+export * from './KebabMenuIcon'
+export * from './LeaderMemberIcon'
 export * from './Loader'
 export * from './NotificationIcon'
 export * from './QuestionIcon'
@@ -17,5 +20,3 @@ export * from './SuccessIcon'
 export * from './TransferIcon'
 export * from './VerifiedMemberIcon'
 export * from './WaitingIcon'
-export * from './ArrowRightIcon'
-export * from './KebabMenuIcon'
