+++ packages/ui/src/memberships/components/MemberRoles.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
-import styled from 'styled-components'
+import styled, { css } from 'styled-components'
 
+import { Colors, Fonts } from '@/common/constants'
 import { memberRoleAbbreviation, memberRoleTitle } from '@/memberships/helpers'
 
-import { MemberRoleHelp, MemberRoleHelpMax, Tooltip } from '../../common/components/Tooltip'
+import { DarkTooltipInnerItemProps, DefaultTooltip, Tooltip, TooltipComponent } from '../../common/components/Tooltip'
 import { MemberRole } from '../types'
 
 interface MemberRolesProps {
@@ -61,6 +62,82 @@ export const MemberRoles = ({ size, max, wrapable, roles }: MemberRolesProps) =>
     </>
   )
 }
+
+interface MemberRoleTooltipProps {
+  size?: 'l' | 'm'
+}
+
+export const MemberRoleHelp = styled(DefaultTooltip)<MemberRoleTooltipProps & DarkTooltipInnerItemProps>`
+  width: ${({ size }) => (size === 'l' ? '24px' : '16px')};
+  height: ${({ size }) => (size === 'l' ? '24px' : '16px')};
+  font-size: ${({ size }) => (size === 'l' ? '10px' : '6px')};
+  line-height: 1;
+  font-family: ${Fonts.Inter};
+  font-weight: 700;
+  ${({ isOnDark }) =>
+    isOnDark
+      ? css`
+          color: ${Colors.Black[300]};
+          background-color: ${Colors.Black[600]};
+          border-color: ${Colors.Black[600]};
+        `
+      : css`
+          color: ${Colors.Black[600]};
+          background-color: ${Colors.Black[100]};
+          border-color: ${Colors.Black[100]};
+        `};
+
+  ${TooltipComponent}:hover > &,
+  ${TooltipComponent}:focus > & {
+    color: ${Colors.White} !important;
+    background-color: ${Colors.Blue[500]} !important;
+    border-color: ${Colors.Blue[500]} !important;
+  }
+`
+
+export const MemberRoleHelpMax = styled(MemberRoleHelp)`
+  background-color: ${Colors.White};
+  color: ${Colors.Blue[500]};
+  border-color: ${Colors.Blue[50]};
+`
+
+export const MemberStatusTooltip = styled(DefaultTooltip)<MemberRoleTooltipProps & DarkTooltipInnerItemProps>`
+  ${({ size }) =>
+    size === 'l'
+      ? css`
+          width: 24px;
+          height: 24px;
+        `
+      : css`
+          width: 16px;
+          height: 16px;
+        `};
+  ${({ isOnDark }) =>
+    isOnDark
+      ? css`
+          border-color: ${Colors.Blue[500]};
+          color: ${Colors.Blue[500]};
+          background-color: transparent;
+        `
+      : css`
+          border-color: ${Colors.Black[200]};
+          color: ${Colors.Black[900]};
+        `};
+
+  ${TooltipComponent}:hover > &,
+  ${TooltipComponent}:focus > & {
+    &.TooltipOnLight {
+      color: ${Colors.Blue[500]};
+      background-color: ${Colors.Black[50]};
+      border-color: ${Colors.Blue[100]};
+    }
+    &.TooltipOnDark {
+      color: ${Colors.Blue[400]} !important;
+      background-color: transparent !important;
+      border-color: ${Colors.Blue[400]} !important;
+    }
+  }
+`
 
 export const MemberRolesWrapperWrapable = styled.div`
   display: grid;
